Rename TorStatusCheckerWrapper to TorStatusCheckerModel
